split up quiz auth tests
