+++ react/test/quiz_auth_1.test.ts
@@ -1,9 +1,7 @@
 import { Selector } from 'testcafe'
 
 import { dissociateUrbanStatsGoogle, email, quizAuthFixture, signInLink, signOutLink, urbanStatsGoogleSignIn } from './auth_test_utils'
-import { addFriend, createUser, restoreUser, startingState } from './quiz_friends_test_utils'
 import { exampleQuizHistory } from './quiz_test_template'
-import { clickButtons } from './quiz_test_utils'
 import { safeReload, target } from './test_utils'
 
 quizAuthFixture('existing state', `${target}/quiz.html`, {
@@ -72,42 +70,4 @@ test('sign in to google, do not enable drive, should not be signed in', async (t
     await t.expect(signInLink.exists).ok()
     await t.expect(signOutLink.exists).notOk()
     await t.expect(Selector('div').withText(`Signed in with ${email}.`).exists).notOk()
-})
-
-quizAuthFixture('no state', `${target}/quiz.html`, {}, '', 'desktop')
-
-test('sync quiz progress two devices', async (t) => {
-    const state = startingState()
-    await createUser(t, 'Alice', '0a', state)
-    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
-    await urbanStatsGoogleSignIn(t)
-    await t.expect(Selector('div').withExactText('1\nPlayed').exists).ok()
-    await createUser(t, 'Bob', '0b', state)
-    await urbanStatsGoogleSignIn(t)
-    await t.navigateTo(`${target}/quiz.html#date=650`)
-    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
-    await t.navigateTo(`${target}/quiz.html`)
-    await t.expect(Selector('div').withExactText('2\nPlayed').exists).ok()
-    await restoreUser(t, 'Alice', state)
-    await t.expect(Selector('div').withExactText('2\nPlayed').exists).ok()
-})
-
-test('sync friends two devices', async (t) => {
-    const state = startingState()
-    await createUser(t, 'Alice', '0a', state)
-    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
-    await addFriend(t, 'Charlie', '0c')
-    await urbanStatsGoogleSignIn(t)
-    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
-    await createUser(t, 'Bob', '0b', state)
-    await t.navigateTo(`${target}/quiz.html`)
-    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
-    await urbanStatsGoogleSignIn(t)
-    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
-    await addFriend(t, 'Darlene', '0d')
-    // give it time to sync
-    await t.wait(5000)
-    await restoreUser(t, 'Alice', state)
-    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
-    await t.expect(Selector('b').withExactText('Darlene').exists).ok()
 })
