+++ react/test/quiz_auth_2.test.ts
@@ -1,0 +1,50 @@
+import { Selector } from 'testcafe'
+
+import { quizAuthFixture, urbanStatsGoogleSignIn } from './auth_test_utils'
+import { addFriend, createUser, removeFriend, restoreUser, startingState } from './quiz_friends_test_utils'
+import { clickButtons } from './quiz_test_utils'
+import { target } from './test_utils'
+
+quizAuthFixture('no state', `${target}/quiz.html`, {}, '', 'desktop')
+
+test('sync quiz progress two devices', async (t) => {
+    const state = startingState()
+    await createUser(t, 'Alice', '0a', state)
+    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
+    await urbanStatsGoogleSignIn(t)
+    await t.expect(Selector('div').withExactText('1\nPlayed').exists).ok()
+    await createUser(t, 'Bob', '0b', state)
+    await urbanStatsGoogleSignIn(t)
+    await t.navigateTo(`${target}/quiz.html#date=650`)
+    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
+    await t.navigateTo(`${target}/quiz.html`)
+    await t.expect(Selector('div').withExactText('2\nPlayed').exists).ok()
+    await restoreUser(t, 'Alice', state)
+    await t.expect(Selector('div').withExactText('2\nPlayed').exists).ok()
+})
+
+test('sync friends two devices', async (t) => {
+    const state = startingState()
+    await createUser(t, 'Alice', '0a', state)
+    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
+    await addFriend(t, 'Charlie', '0c')
+    await urbanStatsGoogleSignIn(t)
+    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
+    await createUser(t, 'Bob', '0b', state)
+    await t.navigateTo(`${target}/quiz.html`)
+    await clickButtons(t, ['a', 'a', 'a', 'a', 'a'])
+    await urbanStatsGoogleSignIn(t)
+    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
+    await addFriend(t, 'Darlene', '0d')
+    // give it time to sync
+    await t.wait(5000)
+    await restoreUser(t, 'Alice', state)
+    await t.expect(Selector('b').withExactText('Charlie').exists).ok()
+    await t.expect(Selector('b').withExactText('Darlene').exists).ok()
+
+    await removeFriend(t, 1)
+    // give it time to sync
+    await t.wait(5000)
+    await restoreUser(t, 'Bob', state)
+    await t.expect(Selector('b').withExactText('Darlene').exists).notOk()
+})
